+++ models/index.js
@@ -12,5 +12,4 @@ Review.belongsTo(User, {
 
 //! Restaurant relationship
 
-
 module.exports = { User, Review };
