+++ public/js/logout.js
