+++ views/home.handlebars
@@ -1,8 +1,5 @@
 <!-- Rubberduckie home page -->
 
-
-
-<!--- old -->
 <h2> Recent Quacks </h2>
 
 <div id="carouselExampleControls" class="carousel" data-bs-ride="carousel">
@@ -15,7 +12,7 @@
                 <div class="card-body">
                     <h5 class="card-title">{{title}}</h5>
                     <p class="card-text">{{content}}</p>
-                    <a href="/api/review/{{id}}" class="view-btn btn btn-dark">View</a>
+                    <a href="/review/{{id}}" class="view-btn btn btn-dark">View</a>
                 </div>
             </div>
         </div>
