+++ views/review.handlebars
@@ -1,14 +1,5 @@
-<h1>{{title}}</h1>
-  <p> {{content}}</p>
-
-<script src=""></script>
-
-{{!-- <div class="container">
-  <h1 class="section-header">{{title}}
-    <span>
-      {{content}}
-    </span>
-  </h1>
+<div class="container">
+  <h1 class="section-header">{{title}}</h1>
   <div class="testimonials">
 
     <div id="carouselExampleControls" class="carousel slide" data-bs-ride="carousel">
@@ -18,115 +9,32 @@
             <div class="row">
               <div class="col-md-5">
                 <div class="profile">
-                  <div class="img-area">
-                    <img src="https://placekitten.com/640/360
-                                                " alt="">
-                  </div>
-                  <div class="radar-area">
-                    <img src="https://static.vecteezy.com/system/resources/thumbnails/003/237/244/small/hud-radar-with-targets-in-action-military-search-system-free-vector.jpg
-                                              " alt="">
-                  </div>
-
+                  {{!-- <div class="img-area">
+                    <img src="https://placekitten.com/640/360" alt="">
+                  </div> --}}
+                  {{!-- <div class="radar-area">
+                    <img src="https://static.vecteezy.com/system/resources/thumbnails/003/237/244/small/hud-radar-with-targets-in-action-military-search-system-free-vector.jpg" alt="">
+                  </div> --}}
                 </div>
               </div>
               <div class="col-md-6">
                 <div class="userPersonal">
-                  <h2> Sage Sun </h2>
-                  <h4>Server</h4>
-                  <h4> 2013-2017 </h4>
+                  <h2>{{username}}</h2>
                 </div>
                 <div class="content">
-                  <p><span><i class="fa fa-quote-left"></i></span>Lorem ipsum dolor sit amet, consectetur
-                    adipisicing elit. Vel a eius excepturi molestias commodi aliquam error magnam consectetur
-                    laboriosam numquam, minima eveniet nostrum sequi saepe ipsam non ea, inventore tenetur!
-                    Corporis commodi consequatur molestiae voluptatum!</p>
-                  <p class="userContact">
+                  <p><span><i class="fa fa-quote-left"></i></span>{{content}}</p>
+                  {{!-- <p class="userContact">
                     <i class="fa fa-twitter"></i>
                     <i class="fa fa-github"></i>
                     <i class="fa fa-linkedin"></i>
-
-                  </p>
-                </div>
-              </div>
-            </div>
-          </div>
-        </div>
-        <div class="carousel-item">
-          <div class="single-item">
-            <div class="row">
-              <div class="col-md-5">
-                <div class="profile">
-                  <div class="img-area">
-                    <img src="https://placeducky.com/single/280/275.gif" alt="">
-                  </div>
-                  <div class="radar-area">
-                    <img src="https://static.vecteezy.com/system/resources/thumbnails/003/237/244/small/hud-radar-with-targets-in-action-military-search-system-free-vector.jpg
-                                              " alt="">
-                  </div>
-
-                </div>
-              </div>
-              <div class="col-md-6">
-                <div class="userPersonal">
-                  <h2>Jeryll Crevinstein</h2>
-                  <h4>Busser</h4>
-                  <h4> 2016-2018 </h4>
-                </div>
-                <div class="content">
-                  <p><span><i class="fa fa-quote-left"></i></span>Lorem ipsum dolor sit amet, consectetur
-                    adipisicing elit. Vel a eius excepturi molestias commodi aliquam error magnam consectetur
-                    laboriosam numquam, minima eveniet nostrum sequi saepe ipsam non ea, inventore tenetur!
-                    Corporis commodi consequatur molestiae voluptatum!</p>
-                  <p class="userContact">
-                    <i class="fa fa-twitter"></i>
-                    <i class="fa fa-github"></i>
-                    <i class="fa fa-linkedin"></i>
-
-                  </p>
-                </div>
-              </div>
-            </div>
-          </div>
-        </div>
-        <div class="carousel-item">
-          <div class="single-item">
-            <div class="row">
-              <div class="col-md-5">
-                <div class="profile">
-                  <div class="img-area">
-                    <img src="https://placeducky.com/real/280/276.png" alt="">
-                  </div>
-                  <div class="radar-area">
-                    <img src="https://static.vecteezy.com/system/resources/thumbnails/003/237/244/small/hud-radar-with-targets-in-action-military-search-system-free-vector.jpg
-                                              " alt="">
-                  </div>
-
-                </div>
-              </div>
-              <div class="col-md-6">
-                <div class="userPersonal">
-                  <h2>Mark Straylo</h2>
-                  <h4>Manager</h4>
-                  <h4> 2012-2019 </h4>
-                </div>
-                <div class="content">
-                  <p><span><i class="fa fa-quote-left"></i></span>Lorem ipsum dolor sit amet, consectetur
-                    adipisicing elit. Vel a eius excepturi molestias commodi aliquam error magnam consectetur
-                    laboriosam numquam, minima eveniet nostrum sequi saepe ipsam non ea, inventore tenetur!
-                    Corporis commodi consequatur molestiae voluptatum!</p>
-                  <p class="userContact">
-                    <i class="fa fa-twitter"></i>
-                    <i class="fa fa-github"></i>
-                    <i class="fa fa-linkedin"></i>
-
-                  </p>
+                  </p> --}}
                 </div>
               </div>
             </div>
           </div>
         </div>
       </div>
-      <button class="carousel-control-prev" type="button" data-bs-target="#carouselExampleControls"
+      {{!-- <button class="carousel-control-prev" type="button" data-bs-target="#carouselExampleControls"
         data-bs-slide="prev">
         <span class="carousel-control-prev-icon" aria-hidden="true"></span>
         <span class="visually-hidden">Previous</span>
@@ -135,12 +43,12 @@
         data-bs-slide="next">
         <span class="carousel-control-next-icon" aria-hidden="true"></span>
         <span class="visually-hidden">Next</span>
-      </button>
+      </button> --}}
     </div>
 
   </div>
 </div>
-</section> --}}
+</section>
 <script src="https://cdnjs.cloudflare.com/ajax/libs/jquery/3.6.0/jquery.min.js"></script>
 <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.2.0/dist/js/bootstrap.bundle.min.js"></script>
 <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.2.2/dist/js/bootstrap.bundle.min.js"
